feat(search): add frecency scoring and improve error handling
- Integrate frecency map into file search results calculation.
- Weight search results with 70% original score and 30% frecency.
- Sort results by the newly calculated combined score.
- Replace `lock().unwrap()` with `lock().map_err(...)` to gracefully handle poisoned mutexes in writer operations.
- Update Cargo.lock dependencies.

+++ launcher-app/src-tauri/src/commands/file_search.rs
@@ -46,7 +46,7 @@ pub fn search_files_by_extension(
         .map_err(|e| format!("Failed to prepare query: {}", e))?;
 
     let results = stmt
-        .query_map(rusqlite::params![ext, limit], |row| {
+        .query_map(rusqlite::params![ext, limit as i64], |row| {
             let frecency: f64 = row.get(6)?;
             Ok(FileSearchResult {
                 id: row.get(0)?,
